# doc/api.yml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: ""
  license:
    name: "MIT"
basePath: "/continents_apiv1"
schemes:
- "https"
paths:
  /:
    get:
      summary: "Lists all continents"
      description: ""
      operationId: "getContinents"
      responses:
        "200":
          description: "Success"
    post:
      summary: "Adds a new continent"
      description: ""
      operationId: "addContinent"
      parameters:
      - name: "continentName"
        in: "query"
        description: "Name of the created continent"
        required: true
        type: "string"

      responses:
        "201":
          description: "Success: URI to the created continent is returned in the Location HTTP header."
        "405":
          description: "Invalid input"
...
  /{continentName}/{countryName}:
    delete:
      summary: "Deletes a country."
      description: ""
      operationId: "deleteCountry"    
      parameters:
      - name: "continentName"
        in: "path"
        description: "Name of the continent where the country is located in"
        required: true
        type: "string"
      - name: "countryName"
        in: "path"
        description: "Name of the country to be deleted."
        required: true
        type: "string"
      responses:
        "200":
          description: "Success"
        "404":
          description: "Country not found"
    get:
      summary: "Retrieves a list of cities in a country"
      description: ""
      operationId: "getCities"
      parameters:
      - name: "continentName"
        in: "path"
        description: "Name of the continent where the country is located in"
        required: true
        type: "string"
      - name: "countryName"
        in: "path"
        description: "Name of the country, which cities are retrieved."
        required: true
        type: "string"
      responses:
        "200":
          description: "Success"
        "404":
          description: "Country not found"
    post:
      summary: "Creates a new city to a country."
      description: ""
      operationId: "AddCity"
      parameters:
      - name: "continentName"
        in: "path"
        description: "Name of the continent where the country is located in"
        required: true
        type: "string"
      - name: "countryName"
        in: "path"
        description: "Name of the country, in which a city is being added."
        required: true
        type: "string"
      responses:
        "201":
          description: "Success: URI to the created city is returned in the Location HTTP header."
        "409":
          description: "City already exists."      
...